Update omniauth-google-oauth2 to 0.6.0
Google announced that they are [shutting down the Google+ APIs in
March][1], which we use through omniauth-google-oauth2 to allow
stuypulse.com logins to TM. The dependent gem released a [new version
0.6.0][2] that fixes the issue.

This should be the only change required of us, without any change to our
OAuth credentials or activated APIs in the Google Developer Console.

This was done with only `bundle update omniauth-google-oauth2`.

[1]: https://developers.google.com/+/api-shutdown
[2]:
https://github.com/zquestz/omniauth-google-oauth2/blob/master/CHANGELOG.md#060---2018-12-28